+++ apiml-security-common/src/main/java/org/zowe/apiml/security/common/login/NonCompulsoryAuthenticationProcessingFilter.java
@@ -1,0 +1,59 @@
+/*
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ */
+
+package org.zowe.apiml.security.common.login;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
+
+import javax.servlet.*;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+/**
+ * Abstract class for filters that can return null Authentication and let the filter chain continue
+ */
+public abstract class NonCompulsoryAuthenticationProcessingFilter extends AbstractAuthenticationProcessingFilter {
+
+    protected NonCompulsoryAuthenticationProcessingFilter(String defaultFilterProcessesUrl) {
+        super(defaultFilterProcessesUrl);
+    }
+
+    @Override
+    public void doFilter(ServletRequest req, ServletResponse res, FilterChain chain) throws IOException, ServletException {
+        HttpServletRequest request = (HttpServletRequest) req;
+        HttpServletResponse response = (HttpServletResponse) res;
+        if (!requiresAuthentication(request, response)) {
+            chain.doFilter(request, response);
+            return;
+        }
+
+        Authentication authResult = null;
+
+        try {
+            authResult = attemptAuthentication(request, response);
+        }
+
+        catch (AuthenticationException failed) {
+            // Authentication failed
+            unsuccessfulAuthentication(request, response, failed);
+            return;
+        }
+
+        if (authResult == null) {
+            chain.doFilter(request, response);
+            return;
+        }
+
+        successfulAuthentication(request, response, chain, authResult);
+    }
+}
